Refacture(Redux,Slice,Main)se hizo el  axios con getAllRooms, se guardo la data con el estado global y se crearon refactorizo la arquitectura del redux
