--- PF-Front/src/redux/roomsActions.js
@@ -1,13 +1,0 @@
-import axios from "axios";
-import { getAllRooms, getRoomById } from "./roomsSlice";
-
-export const getRooms = () => async (dispatch) => {
-    const URL = "mongo.db";
-    try {
-        const dbResponse = await axios(URL);
-        const rooms = dbResponse.data.EXAMPLE;
-        dispatch(getAllRooms(rooms));
-    } catch (error) {
-        console.log(error);
-    }
-};
